% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{CombineAnalytePlots}
\alias{CombineAnalytePlots}
\title{CombineAnalytePlots}
\usage{
CombineAnalytePlots(analytePlots)
}
\arguments{
\item{plot_data}{with columns Analyte, DateTime (Datetime must be POSIXct (format = "%b-%d-%Y")),plot_sym, year, Value, Guideline, minGuideline, maxGuideline,}
}
\value{
facet plot for all analytes, important date (e.g start of mining operations) is shown as a vertical line.
}
\description{
Combines the plots resulting from CreatePlotPerAnalyte
}
