% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{CreatePlotPerAnalyte}
\alias{CreatePlotPerAnalyte}
\title{CreatePlotPerAnalyte}
\usage{
CreatePlotPerAnalyte(gp_plot_data)
}
\value{
list of plots for each analyte
}
\description{
Creates plots for each analyte
}
